# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Documentation for API endpoints",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/contacts/{idContact}/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get list of addresses",
                "operationId": "8e23a407dd7377a40912798006937864",
                "parameters": [
                    {
                        "name": "idContact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List addresses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Create new address",
                "operationId": "e2d24e7ff8b1037f974e619d6eceb5ee",
                "parameters": [
                    {
                        "name": "idContact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "street",
                                    "city",
                                    "province",
                                    "country",
                                    "postal_code"
                                ],
                                "properties": {
                                    "street": {
                                        "type": "string",
                                        "example": "Dharma"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Pandaan"
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "Jawa Timur"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Indonesia"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "67156"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success create address"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contacts/{idContact}/addresses/{idAddress}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get single address",
                "operationId": "8b3727082cc78c6c8dfaa07f24ab90a6",
                "parameters": [
                    {
                        "name": "idContact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idAddress",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address detail"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Update address",
                "operationId": "ce2bf513b2d429062000308419be698b",
                "parameters": [
                    {
                        "name": "idContact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idAddress",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "street": {
                                        "type": "string",
                                        "example": "Dharma Bakti"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Pasuruan"
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "Jatim"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Indonesia"
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "example": "67156"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success update address"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Delete address",
                "operationId": "b3e0563c2c623254eb624732db9a3335",
                "parameters": [
                    {
                        "name": "idContact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idAddress",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success delete address"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}